feat: add 'Show more' button for experiences on mobile
On mobile devices, only the three most recent experiences were displayed. This change introduces a "Show more" button that allows users to view all experiences.

The button is dynamically added to the page if there are more than three experiences. When clicked, it reveals the hidden experiences and then removes itself from the DOM. This improves the user experience on mobile devices by providing access to all content without cluttering the initial view.